// modules/mmorpg/src/model/generated/server/demo/protocol/rpcs.ts
export const MapInstanceProtocol = {
  Resolve: defineRpc<S2L_ResolveMapInstance, L2S_ResolveMapInstance>({
    name: "MapInstance.Resolve",
    requestCode: MsgCode.S2L_ResolveMapInstance,
    responseCode: MsgCode.L2S_ResolveMapInstance,
    requestCodec: S2L_ResolveMapInstanceCodec,
    responseCodec: L2S_ResolveMapInstanceCodec,
  }),
  Register: defineRpc<S2L_RegisterMapInstance, L2S_RegisterMapInstance>({
    name: "MapInstance.Register",
    requestCode: MsgCode.S2L_RegisterMapInstance,
    responseCode: MsgCode.L2S_RegisterMapInstance,
    requestCodec: S2L_RegisterMapInstanceCodec,
    responseCodec: L2S_RegisterMapInstanceCodec,
  }),
  Remove: defineRpc<S2L_RemoveMapInstance, L2S_RemoveMapInstance>({
    name: "MapInstance.Remove",
    requestCode: MsgCode.S2L_RemoveMapInstance,
    responseCode: MsgCode.L2S_RemoveMapInstance,
    requestCodec: S2L_RemoveMapInstanceCodec,
    responseCodec: L2S_RemoveMapInstanceCodec,
  }),
};
// ...
export const LocationProtocol = {
  AllocateUnitId: defineRpc<S2L_AllocatePlayerUnitId, L2S_AllocatePlayerUnitId>({
    name: "Location.AllocateUnitId",
    requestCode: MsgCode.S2L_AllocatePlayerUnitId,
    responseCode: MsgCode.L2S_AllocatePlayerUnitId,
    requestCodec: S2L_AllocatePlayerUnitIdCodec,
    responseCodec: L2S_AllocatePlayerUnitIdCodec,
  }),
  Register: defineRpc<S2L_RegisterPlayerLocation, L2S_RegisterPlayerLocation>({
    name: "Location.Register",
    requestCode: MsgCode.S2L_RegisterPlayerLocation,
    responseCode: MsgCode.L2S_RegisterPlayerLocation,
    requestCodec: S2L_RegisterPlayerLocationCodec,
    responseCodec: L2S_RegisterPlayerLocationCodec,
  }),
  Resolve: defineRpc<S2L_ResolvePlayerLocation, L2S_ResolvePlayerLocation>({
    name: "Location.Resolve",
    requestCode: MsgCode.S2L_ResolvePlayerLocation,
    responseCode: MsgCode.L2S_ResolvePlayerLocation,
    requestCodec: S2L_ResolvePlayerLocationCodec,
    responseCodec: L2S_ResolvePlayerLocationCodec,
  }),
  ResolveMany: defineRpc<S2L_ResolvePlayerLocations, L2S_ResolvePlayerLocations>({
    name: "Location.ResolveMany",
    requestCode: MsgCode.S2L_ResolvePlayerLocations,
    responseCode: MsgCode.L2S_ResolvePlayerLocations,
    requestCodec: S2L_ResolvePlayerLocationsCodec,
    responseCodec: L2S_ResolvePlayerLocationsCodec,
  }),
  Lock: defineRpc<S2L_LockPlayerLocation, L2S_LockPlayerLocation>({
    name: "Location.Lock",
    requestCode: MsgCode.S2L_LockPlayerLocation,
    responseCode: MsgCode.L2S_LockPlayerLocation,
    requestCodec: S2L_LockPlayerLocationCodec,
    responseCodec: L2S_LockPlayerLocationCodec,
  }),
  Commit: defineRpc<S2L_CommitPlayerLocation, L2S_CommitPlayerLocation>({
    name: "Location.Commit",
    requestCode: MsgCode.S2L_CommitPlayerLocation,
    responseCode: MsgCode.L2S_CommitPlayerLocation,
    requestCodec: S2L_CommitPlayerLocationCodec,
    responseCodec: L2S_CommitPlayerLocationCodec,
  }),
  RebindGate: defineRpc<S2L_RebindPlayerGate, L2S_RebindPlayerGate>({
    name: "Location.RebindGate",
    requestCode: MsgCode.S2L_RebindPlayerGate,
    responseCode: MsgCode.L2S_RebindPlayerGate,
    requestCodec: S2L_RebindPlayerGateCodec,
    responseCodec: L2S_RebindPlayerGateCodec,
  }),
  Unlock: defineRpc<S2L_UnlockPlayerLocation, L2S_UnlockPlayerLocation>({
    name: "Location.Unlock",
    requestCode: MsgCode.S2L_UnlockPlayerLocation,
    responseCode: MsgCode.L2S_UnlockPlayerLocation,
    requestCodec: S2L_UnlockPlayerLocationCodec,
    responseCodec: L2S_UnlockPlayerLocationCodec,
  }),
  Remove: defineRpc<S2L_RemovePlayerLocation, L2S_RemovePlayerLocation>({
    name: "Location.Remove",
    requestCode: MsgCode.S2L_RemovePlayerLocation,
    responseCode: MsgCode.L2S_RemovePlayerLocation,
    requestCodec: S2L_RemovePlayerLocationCodec,
    responseCodec: L2S_RemovePlayerLocationCodec,
  }),
  RecoverOwner: defineRpc<S2L_RecoverPlayerLocations, L2S_RecoverPlayerLocations>({
    name: "Location.RecoverOwner",
    requestCode: MsgCode.S2L_RecoverPlayerLocations,
    responseCode: MsgCode.L2S_RecoverPlayerLocations,
    requestCodec: S2L_RecoverPlayerLocationsCodec,
    responseCodec: L2S_RecoverPlayerLocationsCodec,
  }),
};